Adding Norification Functionality

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -63,19 +63,47 @@ export async function POST(
             return Response.json({error: "Unauthorized"}, {status: 401})
         }
 
-        await prisma.like.upsert({
+        const post = await prisma.post.findUnique({
             where : {
-                userId_postId : {
+                id: postId
+            },
+            select: {
+                userId: true
+            }
+        })
+
+        if(!post){
+            return Response.json({error: "Post not found"}, {status: 404})
+        }
+
+        await prisma.$transaction([
+            prisma.like.upsert({
+                where : {
+                    userId_postId : {
+                        userId : loggedInUser.id,
+                        postId : postId
+                    }
+                },
+                create : {
                     userId : loggedInUser.id,
                     postId : postId
-                }
-            },
-            create : {
-                userId : loggedInUser.id,
-                postId : postId
-            },
-            update : {}
-        })
+                },
+                update : {}
+            }),
+            ...(loggedInUser.id !== post.userId) ?
+            [
+                prisma.notification.create({
+                    data : {
+                        issuerId: loggedInUser.id,
+                        recipientId: post.userId,
+                        postId : postId,
+                        type : "LIKE"
+                    }
+                })
+            ]
+            :
+            []
+        ])
 
         return new Response();
     } catch (error) {
@@ -96,12 +124,35 @@ export async function DELETE(
             return Response.json({error: "Unauthorized"}, {status: 401})
         }
 
-        await prisma.like.deleteMany({
+        const post = await prisma.post.findUnique({
             where : {
-                userId : loggedInUser.id,
-                postId : postId       
+                id: postId
+            },
+            select: {
+                userId: true
             }
         })
+
+        if(!post){
+            return Response.json({error: "Post not found"}, {status: 404})
+        }
+
+        await prisma.$transaction([
+            prisma.like.deleteMany({
+                where : {
+                    userId : loggedInUser.id,
+                    postId : postId       
+                }
+            }),
+            prisma.notification.deleteMany({
+                where : {
+                    issuerId: loggedInUser.id,
+                    recipientId: post.userId,
+                    postId : postId,
+                    type : "LIKE"
+                }
+            })
+        ])
 
         return new Response();
     } catch (error) {
